fluentvalidation, move detectorlogic
fluentvalidation,  move detection logic to tree engine
