+++ WhisperBranch.Contracts/Interfaces/INode.cs
@@ -1,0 +1,9 @@
+﻿namespace WhisperBranch.Contracts.Interfaces
+{
+    public interface INode
+    {
+        public int Id { get; set; }
+
+        public IEnumerable<INode> Children { get; set; }
+    }
+}
